# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ki-db Knowledge Index API
  version: 1.0.0
  description: |
    Postgres-centric Knowledge Index DB with hybrid BM25+vector search,
    knowledge graph, and event-driven indexing pipeline.
servers:
  - url: http://localhost:8080
    description: Ingest API
  - url: http://localhost:8081
    description: Query API
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
  schemas:
    Document:
      type: object
      required: [doc_id, stable_key, doc_type, status]
      properties:
        tenant_id:
          type: string
          format: uuid
        doc_id:
          type: string
          format: uuid
        stable_key:
          type: string
        title:
          type: string
        doc_type:
          type: string
          enum: [report, adr, postmortem, snippet, glossary, guide, policy, other]
        status:
          type: string
          enum: [inbox, published, deprecated, archived]
        confidence:
          type: string
          enum: [high, med, low]
        owners:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Edge:
      type: object
      properties:
        from_doc_id:
          type: string
          format: uuid
        to_doc_id:
          type: string
          format: uuid
        to_external_key:
          type: string
        relation:
          type: string
          enum: [links_to, references, supersedes, duplicates, contradicts]
...
security:
  - TenantHeader: []
paths:
...
  /api/v1/graph/ego:
    get:
      tags: [Graph]
      summary: Get ego graph around a document
      operationId: getEgoGraph
      parameters:
        - name: doc_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: hops
          in: query
          schema:
            type: integer
            default: 2
      responses:
        '200':
          description: Graph nodes and edges
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  edges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Edge'
...